refactor(policy): 抽出 POLICY_OPTIONS 與 toPolicyReviewInput 供制度表單共用
