+++ src/components/ShareExperience/PolicyForm/toPolicyReviewInput.ts
@@ -1,0 +1,69 @@
+import { PolicyReviewInput } from 'apis/createPolicyReviewGroup';
+
+const policyEnumMap: Record<string, string> = {
+  生理假: 'MENSTRUAL_LEAVE',
+  育嬰假: 'PARENTAL_LEAVE',
+  家庭照顧假: 'FAMILY_CARE_LEAVE',
+  彈性上下班時間: 'FLEXIBLE_WORKING_HOUR',
+  遠端工作: 'REMOTE_WORK',
+};
+
+const hasPolicyMap: Record<string, string> = {
+  是: 'yes',
+  有: 'yes',
+  否: 'no',
+  沒有: 'no',
+  不知道: 'unknown',
+};
+
+const complianceMap: Record<string, string> = {
+  '有，優於性別平等工作法': 'yes',
+  '有，符合性別平等工作法': 'yes',
+  '有，不符合性別平等工作法': 'no',
+  '有，不清楚是否符合性別平等工作法': 'unknown',
+};
+
+const remoteWorkPolicyMap: Record<string, string> = {
+  每週一天: 'ONE_DAY_PER_WEEK',
+  每週兩天: 'TWO_DAYS_PER_WEEK',
+  每週三天: 'THREE_DAYS_PER_WEEK',
+  每週四天: 'FOUR_DAYS_PER_WEEK',
+  不限天數: 'NO_LIMIT',
+};
+
+const toPolicyReviewInput = ([
+  optionValue,
+  radioValue,
+  elseOptionValue,
+  textValue,
+]: unknown[]): PolicyReviewInput => {
+  const policy = policyEnumMap[optionValue as string];
+  const hasPolicy = hasPolicyMap[radioValue as string];
+  const review = (textValue as string) || undefined;
+
+  if (policy === 'REMOTE_WORK') {
+    return {
+      policy,
+      hasPolicy,
+      review,
+      remoteWorkPolicy: elseOptionValue
+        ? remoteWorkPolicyMap[elseOptionValue as string]
+        : undefined,
+    };
+  }
+
+  if (policy === 'FLEXIBLE_WORKING_HOUR') {
+    return { policy, hasPolicy, review };
+  }
+
+  return {
+    policy,
+    hasPolicy,
+    review,
+    compliance: elseOptionValue
+      ? complianceMap[elseOptionValue as string]
+      : undefined,
+  };
+};
+
+export default toPolicyReviewInput;
